requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
chromadb
sentence-transformers
huggingface-hub
google-genai
python-dotenv
pydantic
